edit activity server integration
