+++ app/models/api/axios-api/activity.ts
@@ -1,0 +1,6 @@
+import { UserActivities } from "../../responses/axios/user";
+import { Response } from "../../responses/axios/response";
+
+export interface activityApi {
+    getActivities(): Promise<Response<UserActivities>>;
+}
