+++ app/models/api/axios-api/expense.ts
@@ -8,7 +8,6 @@ import {
   UserExpenses,
 } from "../../responses/axios/user";
 import { Response } from "../../responses/axios/response";
-import { Token } from "../../other/axios/Token";
 
 export interface expenseApi {
   getExpenses(): Promise<Response<UserExpenses>>;
