+++ app/models/requests/axios/user.ts
@@ -90,19 +90,3 @@ export interface GetExpenseCommentsRequest {
 export interface GetUserActivitiesRequest {
   token: Token;
 }
-
-export interface GetUserActivityRequest {
-  token: Token;
-}
-
-export interface AddActivityRequest {
-  token: Token;
-}
-
-export interface DeleteActivityRequest {
-  token: Token;
-}
-
-export interface EditActivityRequest {
-  token: Token;
-}
