+++ app/store/actions/expenseActions.ts
@@ -9,11 +9,6 @@ import {
   GetExpenseCommentsRequest,
   EditExpenseRequest,
   DeleteExpenseRequest,
-  GetUserActivitiesRequest,
-  AddActivityRequest,
-  DeleteActivityRequest,
-  GetUserActivityRequest,
-  EditActivityRequest,
 } from "../../models/requests/axios/user";
 import {
   UserExpenses,
@@ -21,179 +16,9 @@ import {
   UserExpense,
   ExpenseComments,
   EditExpense,
-  UserActivities,
-  UserActivity,
-  AddActivity,
-  DeleteActivity,
-  EditActivity,
 } from "../../models/responses/axios/user";
 import { Response } from "../../models/responses/axios/response";
 import * as types from "./types";
-
-// Get activities
-export function onGetUserActivitiesRequest(token: Token): Action<GetUserActivitiesRequest> {
-  return {
-    type: types.GET_USER_ACTIVITIES_REQUEST,
-    payload: {
-      token: token,
-    }
-  };
-}
-
-export function onGetUserActivitiesResponse(
-  response: Response<UserActivities>
-): Action<Response<UserActivities>> {
-  return {
-    type: types.GET_USER_ACTIVITIES_RESPONSE,
-    payload: {
-      success: response.success,
-      status: response.status,
-      response: response.response,
-    },
-  };
-}
-
-export function onGetUserActivitiesFail(): Action<Response<UserActivities>> {
-  return {
-    type: types.GET_USER_ACTIVITIES_FAIL,
-    payload: {
-      success: false,
-      status: -1,
-    },
-  };
-}
-
-// Get activity
-export function onGetUserActivityRequest(token: Token): Action<GetUserActivityRequest> {
-  return {
-    type: types.GET_USER_ACTIVITY_REQUEST,
-    payload: {
-      token: token,
-    }
-  };
-}
-
-export function onGetUserActivityResponse(
-  response: Response<UserActivity>
-): Action<Response<UserActivity>> {
-  return {
-    type: types.GET_USER_ACTIVITY_RESPONSE,
-    payload: {
-      success: response.success,
-      status: response.status,
-      response: response.response,
-    },
-  };
-}
-
-export function onGetUserActivityFail(): Action<Response<UserActivity>> {
-  return {
-    type: types.GET_USER_ACTIVITY_FAIL,
-    payload: {
-      success: false,
-      status: -1,
-    },
-  };
-}
-
-// Add Activity
-export function onAddActivityRequest(token: Token): Action<AddActivityRequest> {
-  return {
-    type: types.ADD_ACTIVITY_REQUEST,
-    payload: {
-      token: token,
-    }
-  };
-}
-
-export function onAddActivityResponse(
-  response: Response<AddActivity>
-): Action<Response<AddActivity>> {
-  return {
-    type: types.ADD_ACTIVITY_RESPONSE,
-    payload: {
-      success: response.success,
-      status: response.status,
-      response: response.response,
-    },
-  };
-}
-
-export function onAddActivityFail(): Action<Response<AddActivity>> {
-  return {
-    type: types.ADD_ACTIVITY_FAIL,
-    payload: {
-      success: false,
-      status: -1,
-    },
-  };
-}
-
-// Delete activity
-export function onDeleteActivityRequest(token: Token): Action<DeleteActivityRequest> {
-  return {
-    type: types.DELETE_ACTIVITY_REQUEST,
-    payload: {
-      token: token,
-    }
-  };
-}
-
-export function onDeleteActivityResponse(
-  response: Response<DeleteActivity>
-): Action<Response<DeleteActivity>> {
-  return {
-    type: types.DELETE_ACTIVITY_RESPONSE,
-    payload: {
-      success: response.success,
-      status: response.status,
-      response: response.response,
-    },
-  };
-}
-
-export function onDeleteActivityFail(): Action<Response<DeleteActivity>> {
-  return {
-    type: types.DELETE_ACTIVITY_FAIL,
-    payload: {
-      success: false,
-      status: -1,
-    },
-  };
-}
-
-// Edit activity
-export function onEditActivityRequest(token: Token): Action<EditActivityRequest> {
-  return {
-    type: types.EDIT_ACTIVITY_REQUEST,
-    payload: {
-      token: token,
-    }
-  };
-}
-
-export function onEditActivityResponse(
-  response: Response<EditActivity>
-): Action<Response<EditActivity>> {
-  return {
-    type: types.EDIT_ACTIVITY_RESPONSE,
-    payload: {
-      success: response.success,
-      status: response.status,
-      response: response.response,
-    },
-  };
-}
-
-export function onEditActivityFail(): Action<Response<EditActivity>> {
-  return {
-    type: types.EDIT_ACTIVITY_FAIL,
-    payload: {
-      success: false,
-      status: -1,
-    },
-  };
-}
 
 export function onGetUserExpensesRequest(token: Token): Action<GetUserExpensesRequest> {
   return {
