+++ app/store/reducers/userReducer.ts
@@ -370,66 +370,6 @@ export const userReducer = createReducer(initialState, {
   [types.GET_USER_ACTIVITIES_FAIL](state: IUserState, action: Action<Response<UserActivities>>): IUserState {
     return state;
   },
-  [types.GET_USER_ACTIVITY_REQUEST](state: IUserState, action: Action<GetUserActivityRequest>): IUserState {
-    return {
-      ...state,
-      token: action.payload.token,
-    };
-  },
-  [types.GET_USER_ACTIVITY_RESPONSE](
-    state: IUserState,
-    action: Action<Response<UserActivity>>
-  ): IUserState {
-    return state;
-  },
-  [types.GET_USER_ACTIVITY_FAIL](state: IUserState, action: Action<Response<UserActivity>>): IUserState {
-    return state;
-  },
-  [types.ADD_ACTIVITY_REQUEST](state: IUserState, action: Action<AddActivityRequest>): IUserState {
-    return {
-      ...state,
-      token: action.payload.token,
-    };
-  },
-  [types.ADD_ACTIVITY_RESPONSE](
-    state: IUserState,
-    action: Action<Response<AddActivity>>
-  ): IUserState {
-    return state;
-  },
-  [types.ADD_ACTIVITY_FAIL](state: IUserState, action: Action<Response<AddActivity>>): IUserState {
-    return state;
-  },
-  [types.DELETE_ACTIVITY_REQUEST](state: IUserState, action: Action<DeleteActivityRequest>): IUserState {
-    return {
-      ...state,
-      token: action.payload.token,
-    };
-  },
-  [types.DELETE_ACTIVITY_RESPONSE](
-    state: IUserState,
-    action: Action<Response<DeleteActivity>>
-  ): IUserState {
-    return state;
-  },
-  [types.DELETE_ACTIVITY_FAIL](state: IUserState, action: Action<Response<DeleteActivity>>): IUserState {
-    return state;
-  },
-  [types.EDIT_ACTIVITY_REQUEST](state: IUserState, action: Action<EditActivityRequest>): IUserState {
-    return {
-      ...state,
-      token: action.payload.token,
-    };
-  },
-  [types.EDIT_ACTIVITY_RESPONSE](
-    state: IUserState,
-    action: Action<Response<EditActivity>>
-  ): IUserState {
-    return state;
-  },
-  [types.EDIT_ACTIVITY_FAIL](state: IUserState, action: Action<Response<EditActivity>>): IUserState {
-    return state;
-  },
   [types.ADD_EXPENSE_REQUEST](state: IUserState, action: Action<AddExpenseRequest>): IUserState {
     return {
       ...state,
